feat: 增加节点切换

+++ agent/server/server.go
@@ -6,6 +6,7 @@ import (
 	"os"
 
 	"github.com/1Panel-dev/1Panel/agent/cron"
+	"github.com/1Panel-dev/1Panel/agent/global"
 	"github.com/1Panel-dev/1Panel/agent/i18n"
 	"github.com/1Panel-dev/1Panel/agent/init/app"
 	"github.com/1Panel-dev/1Panel/agent/init/business"
@@ -39,11 +40,25 @@ func Start() {
 	server := &http.Server{
 		Handler: rootRouter,
 	}
-
-	_ = os.Remove("/tmp/agent.sock")
-	listener, err := net.Listen("unix", "/tmp/agent.sock")
-	if err != nil {
-		panic(err)
+	if len(global.CurrentNode) == 0 || global.CurrentNode == "127.0.0.1" {
+		_ = os.Remove("/tmp/agent.sock")
+		listener, err := net.Listen("unix", "/tmp/agent.sock")
+		if err != nil {
+			panic(err)
+		}
+		_ = server.Serve(listener)
+	} else {
+		server.Addr = "0.0.0.0:9999"
+		type tcpKeepAliveListener struct {
+			*net.TCPListener
+		}
+		ln, err := net.Listen("tcp4", "0.0.0.0:9999")
+		if err != nil {
+			panic(err)
+		}
+		global.LOG.Info("listen at http://0.0.0.0:9999")
+		if err := server.Serve(tcpKeepAliveListener{ln.(*net.TCPListener)}); err != nil {
+			panic(err)
+		}
 	}
-	_ = server.Serve(listener)
 }
